Posted latest changes from trello
changed all files from .html to .php to incorporate the header and footer files for easier development. The javascript might be reacting to the change and not displaying pictures on the main splash for the vehicle slider. Might to revert to before I started working to see if the slider was working at that point. It is working on the old dev site with the coldbrew subdomain.